refactoring edit test with exception edit test with output to the terminal add test_product_add_exception add abstract class BaseProduct for Product
