+++ src/product/base_product.py
@@ -1,0 +1,5 @@
+from abc import ABC, abstractmethod
+
+
+class BaseProduct(ABC):
+    pass
