+++ src/product/product.py
@@ -1,4 +1,7 @@
-class Product:
+from src.product.base_product import BaseProduct
+
+
+class Product(BaseProduct):
     name: str
     description: str
     __price: float
@@ -47,28 +50,3 @@ class Product:
 
             return cls(**new_product)
 
-
-class Smartphone(Product):
-    efficiency: float
-    model: str
-    memory: int
-    color: str
-
-    def __init__(self, name, description, price, quantity, efficiency, model, memory, color):
-        super().__init__(name, description, price, quantity)
-        self.efficiency = efficiency
-        self.model = model
-        self.memory = memory
-        self.color = color
-
-
-class LawnGrass(Product):
-    country: str
-    germination_period: str
-    color: str
-
-    def __init__(self, name, description, price, quantity, country, germination_period, color):
-        super().__init__(name, description, price, quantity)
-        self.country = country
-        self.germination_period = germination_period
-        self.color = color
